requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.3
